pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-tm1"
version = "0.0.10"
authors = [
  {name = "Akos Andras Nagy", email = "akos.nagy@knowledgeseed.ch"},
  {name = "Marius Wirtz", email = "mwirtz@cubewise.com"}
]
maintainers = [
  {name = "Akos Andras Nagy", email = "akos.nagy@knowledgeseed.ch"},
  {name = "Marius Wirtz", email = "mwirtz@cubewise.com"}
]
description = "A package to simplify connecting to the TM1 REST API from Apache Airflow"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
  "apache-airflow >= 2.3",
  "tm1py >= 1.1"
]

[project.urls]
Documentation = "https://github.com/airflow-provider-tm1/airflow-provider-tm1/blob/master/README.md"
Repository = "https://github.com/airflow-provider-tm1/airflow-provider-tm1"
Issues = "https://github.com/airflow-provider-tm1/airflow-provider-tm1/issues"
Changelog = "https://github.com/airflow-provider-tm1/airflow-provider-tm1/CHANGELOG.md"

[tool.setuptools.packages]
find = {namespaces = false} 

[project.entry-points."apache_airflow_provider"]
provider_info = "airflow_provider_tm1.__init__:get_provider_info"


[tool.black]
line-length = 120
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | _build
    | buck-out
    | build
    | dist
  )
)
'''

[tool.isort]
profile = "black"
